libnetwork: merge linux-only test-files
